rename functions.js directory
