+++ src/scripts/functions.js
